pyproject.toml
[project]
name = "fediway"
version = "0.1.0"
description = "Algorithmic feeds for Mastodon"
readme = "README.md"
license = { text = "AGPL-3.0" }
requires-python = ">=3.10"
authors = [{ name = "Fediway" }]
keywords = ["mastodon", "fediverse", "recommendations", "algorithmic-feed"]

dependencies = [
    # Web framework
    "fastapi[standard]>=0.115",
    "uvicorn>=0.34",
    # Database
    "psycopg2-binary>=2.9",
    "sqlmodel>=0.0.24",
    "alembic>=1.15",
    # Redis
    "redis>=4.5",
    # CLI
    "typer>=0.15",
    # Logging
    "loguru>=0.7",
    # Config
    "pydantic-settings>=2.8",
    # Background tasks
    "celery>=5.4",
    # Utilities
    "tqdm>=4.67",
    "numpy>=1.26",
    "pandas>=2.0",
]

[project.optional-dependencies]
vectors = [
    "qdrant-client>=1.16",
]
embeddings = [
    "sentence-transformers[onnx]>=4.1",
    "multilingual-clip>=1.0",
]
streaming = [
    "faststream>=0.5",
    "confluent-kafka>=2.9",
    "kafka-python>=2.0",
]
ml = [
    "scikit-learn>=1.6",
    "lightgbm>=4.6",
]
features = [
    "feast[redis,postgres]>=0.48",
]
geo = [
    "maxminddb>=2.6",
]
aws = [
    "aiobotocore>=2.17",
    "boto3>=1.35",
]
orbit = [
    "networkx>=3.0",
]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-cov>=7.0",
    "ruff>=0.9",
    "pre-commit>=4.0",
    "fediway[vectors,streaming]",
]
# Convenience groups
full = [
    "fediway[vectors,embeddings,streaming,ml,features,geo,aws,orbit]",
]

[project.urls]
Homepage = "https://github.com/fediway/fediway"
Documentation = "https://about.fediway.com"

[project.scripts]
fediway = "apps.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config", "modules", "shared"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Re-exports are intentional

[tool.ruff.lint.isort]
known-first-party = ["apps", "config", "modules", "shared"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
